Migrate @syncit/ui to rrweb 2.0
