--- packages/ui/public/app.html
@@ -1,30 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>Document</title>
-    <link rel="stylesheet" href="./build/bundle.css" />
-    <style>
-      html,
-      body {
-        height: 100%;
-        width: 100%;
-      }
-    </style>
-  </head>
-  <body>
-    <script src="https://cdn.jsdelivr.net/npm/@syncit/transporter@0.3.2/dist/index.js"></script>
-    <script src="./build/bundle.js"></script>
-    <script>
-      new syncit.App({
-        target: document.body,
-        props: {
-          createTransporter({ role, uid }) {
-            return new syncitTransporter.LocalStorageTransporter();
-          },
-        },
-      });
-    </script>
-  </body>
-</html>
