--- packages/ui/public/embed.html
@@ -1,74 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="utf-8" />
-    <meta
-      name="viewport"
-      content="width=device-width, initial-scale=1, shrink-to-fit=no"
-    />
-    <meta name="description" content="" />
-    <meta name="author" content="" />
-
-    <title>Signin Template for Bootstrap</title>
-
-    <link
-      rel="canonical"
-      href="https://getbootstrap.com/docs/4.0/examples/sign-in/"
-    />
-
-    <link rel="stylesheet" href="./build/bundle.css" />
-    <!-- Bootstrap core CSS -->
-    <link href="./embed.css" rel="stylesheet" />
-  </head>
-
-  <body class="text-center">
-    <form class="form-signin">
-      <img
-        class="mb-4"
-        src="https://getbootstrap.com/docs/4.0/assets/brand/bootstrap-solid.svg"
-        alt=""
-        width="72"
-        height="72"
-      />
-      <h1 class="h3 mb-3 font-weight-normal">Please sign in</h1>
-      <label for="inputEmail" class="sr-only">Email address</label>
-      <input
-        type="email"
-        id="inputEmail"
-        class="form-control"
-        placeholder="Email address"
-        required
-        autofocus
-      />
-      <label for="inputPassword" class="sr-only">Password</label>
-      <input
-        type="password"
-        id="inputPassword"
-        class="form-control"
-        placeholder="Password"
-        required
-      />
-      <div class="checkbox mb-3">
-        <label>
-          <input type="checkbox" value="remember-me" /> Remember me
-        </label>
-      </div>
-      <button class="btn btn-lg btn-primary btn-block" type="submit">
-        Sign in
-      </button>
-      <p class="mt-5 mb-3 text-muted">&copy; 2017-2018</p>
-    </form>
-    <script src="https://cdn.jsdelivr.net/npm/@syncit/transporter@0.3.2/dist/index.js"></script>
-    <script src="./build/bundle.js"></script>
-    <script>
-      new syncit.Embed({
-        target: document.body,
-        props: {
-          createTransporter({ role, uid }) {
-            return new syncitTransporter.LocalStorageTransporter();
-          },
-        },
-      });
-    </script>
-  </body>
-</html>
