+++ packages/ui/src/lib/components/Icon.svelte
