+++ packages/ui/src/lib/components/LineChart.svelte
