+++ packages/ui/src/lib/components/Tag.svelte.d.ts
@@ -1,0 +1,35 @@
+/** @typedef {typeof __propDef.props}  TagProps */
+/** @typedef {typeof __propDef.events}  TagEvents */
+/** @typedef {typeof __propDef.slots}  TagSlots */
+export default class Tag extends SvelteComponentTyped<{
+    [x: string]: never;
+}, {
+    mouseover: MouseEvent;
+    mouseout: MouseEvent;
+    click: MouseEvent;
+} & {
+    [evt: string]: CustomEvent<any>;
+}, {
+    default: {};
+}> {
+}
+export type TagProps = typeof __propDef.props;
+export type TagEvents = typeof __propDef.events;
+export type TagSlots = typeof __propDef.slots;
+import { SvelteComponentTyped } from "svelte";
+declare const __propDef: {
+    props: {
+        [x: string]: never;
+    };
+    events: {
+        mouseover: MouseEvent;
+        mouseout: MouseEvent;
+        click: MouseEvent;
+    } & {
+        [evt: string]: CustomEvent<any>;
+    };
+    slots: {
+        default: {};
+    };
+};
+export {};
