+++ packages/ui/src/lib/index.ts
@@ -1,0 +1,8 @@
+// Reexport your entry components here
+export { default as App } from './App.svelte';
+export { default as Embed } from './Embed.svelte';
+
+// import App from '../App.svelte';
+// import Embed from '../Embed.svelte';
+
+// export { App, Embed };
