+++ packages/ui/src/locales/enUS.d.ts
@@ -1,0 +1,43 @@
+export namespace locale {
+    const lng: string;
+    namespace data {
+        namespace app {
+            const connect: string;
+            const initializing: string;
+            const ready: string;
+            const latency: string;
+            const bandwidth: string;
+            const remoteControl: string;
+            const requestToControl: string;
+            const requested: string;
+            const stopControl: string;
+            const aborted: string;
+            const reset: string;
+        }
+        namespace embed {
+            const initializing_1: string;
+            export { initializing_1 as initializing };
+            export const cancel: string;
+            export const selectBlockArea: string;
+            export const area: string;
+            export const enableShare: string;
+            const ready_1: string;
+            export { ready_1 as ready };
+            export const mouseSize: string;
+            export const small: string;
+            export const medium: string;
+            export const large: string;
+            const remoteControl_1: string;
+            export { remoteControl_1 as remoteControl };
+            export const disabled: string;
+            export const requesting: string;
+            export const enabled: string;
+            export const accept: string;
+            export const abortControl: string;
+            export const abort: string;
+            export const copied: string;
+            export const paint: string;
+            export const stopPaint: string;
+        }
+    }
+}
