+++ packages/ui/src/locales/index.d.ts
@@ -1,0 +1,2 @@
+export function t(path: any, replacement: any): any;
+export function setCurrentLanguage(lng: any): void;
