--- packages/ui/src/main.js
@@ -1,4 +1,0 @@
-import App from './App.svelte';
-import Embed from './Embed.svelte';
-
-export { App, Embed };
