+++ packages/ui/static/embed.css
@@ -8313,7 +8313,7 @@ a.text-dark:hover {
     border: 1px solid #ddd !important;
   }
 }
-/*# sourceMappingURL=bootstrap.min.css.map */
+
 
 html,
 body {
